backend/requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
sentence-transformers
openai
langchain-openai
langchain-community
python-dotenv
chromadb
